requirements.txt
alembic==1.13.1
bcrypt==4.1.3
blinker==1.8.2
click==8.1.7
dnspython==2.6.1
email_validator==2.1.1
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.0.3
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==2.1.5
pillow==10.3.0
SQLAlchemy==2.0.30
typing_extensions==4.12.1
Werkzeug==3.0.3
WTForms==3.1.2